Added functions to summarize subsamplings
User supplied functions can be used to summarize subsamplings
and calculate q-values. Previously, both estimation of pi0 and
computation of summary statistics were based on
the subsample with the highest proportion of observed reads.
This leads to some unintuitive behavior if someone tries to split up
computation among several calls to subsample(). The new method allows
the user to specify a function that calculates q-values and summaries for
a subsample object.